added scss, fixed bem
