+++ src/css/mixins.scss
@@ -1,0 +1,5 @@
+@mixin flexColumnStart {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+}
