+++ src/css/style.scss
@@ -1,0 +1,69 @@
+@import '../css/vars.scss';
+@import '../css/mixins.scss';
+@import '../css/header.scss';
+@import '../css/nav.scss';
+@import '../css/hot_keys.scss';
+@import '../css/tabs.scss';
+@import '../css/card.scss';
+@import '../css/footer.scss';
+
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+    font-family: 'Open Sans', sans-serif;
+}
+
+body {
+    width: 100%;
+    height: 100vh;
+    display: grid;
+    grid-template-rows: auto auto;
+}
+
+.wrapper {
+    width: 100%;
+    display: grid;
+    justify-content: center;
+    grid-template-rows: 50px 1fr;
+    padding-top: 32px;
+}
+
+.header {
+    width: 1368px;
+    height: 48px;
+    margin: 0 40px 0 32px;
+    display: grid;
+    grid-template-columns: 280px 894px auto;
+}
+
+.main {
+    padding-left: 20px;
+    margin-top: 30px;
+    margin-right: 88px;
+    display: flex;
+
+    &__content {
+        width: 1038px;
+        display: flex;
+        flex-direction: column;
+    }
+
+    &__items {
+        display: grid;
+        grid-template-columns: repeat(auto-fill, minmax(330px, 1fr));
+        gap: 24px;
+    }
+}
+
+footer {
+    width: 100%;
+    height: fit-content;
+    background: $black;
+    display: flex;
+    justify-content: space-between;
+    padding: 32px;
+    align-items: center;
+    margin-top: 50px;
+    align-self: end;
+}
