+++ src/css/vars.scss
@@ -1,0 +1,6 @@
+$black: #1c1819;
+$white: #fff;
+$blue: #2a81dd;
+$red: #f33333;
+$shadow-color: rgba(28, 24, 25, 0.1);
+$radius: 8px;
